+++ CHANGES.md
@@ -1,4 +1,4 @@
-## 0.1.0-alpha2
+## 0.1.0-alpha2 [source](https://github.com/gadfly361/soda-ash/tree/4f266c3d3a9dbca240238a3c6f3d2f62cf80586a)
 
 * initial release
 * targets Semantic UI version 2.1.8 and Reagent 0.5.1